+++ src/Game/Text/HelpText.en.sjson.lua
@@ -7,14 +7,31 @@ local order = {
 local newData = {
   {
     Id = "Song_RandomSong_PlayingInfo",
-    DisplayName = "Music Maker's choice",
+    DisplayName = "Music Maker's Choice",
     Description =
-    "Allow the Music Maker to choose a new song for you each time you return to the crossroads.\n Now playing {#ItalicFormat}{$TempTextData.PlayingSongFriendlyName}{#prev}."
+    "Allow the Music Maker to choose a new song for you each time you return to the crossroads.\n Now playing {#ItalicFormat}{$TempTextData.PlayingSongFriendlyName}."
   },
   {
     Id = "Song_RandomSong",
-    DisplayName = "Music Maker's choice",
+    DisplayName = "Music Maker's Choice",
     Description = "Allow the Music Maker to choose a new song for you each time you return to the crossroads."
+  },
+  {
+    Id = "Song_RandomSongFavorites_PlayingInfo",
+    DisplayName = "Music Maker's Choice",
+    Description =
+    "Allow the Music Maker to choose a new, favorited song for you each time you return to the crossroads.\n Now playing {#ItalicFormat}{$TempTextData.PlayingSongFriendlyName}."
+  },
+  {
+    Id = "Song_RandomSongFavorites",
+    DisplayName = "Music Maker's Choice (Favorites)",
+    Description =
+    "Allow the Music Maker to choose a new, favorited song for you each time you return to the crossroads.\n {#ItalicFormat}Favorite songs using   {IP}."
+  },
+  -- Shows up as a resource cost when trying to buy the random favorite song
+  {
+    Id = "ModsNikkelMMusicMakerRandomizerMusicPlayerFavoritesCount_Short",
+    DisplayName = "At least {#ItalicFormat}one{#prev}\nfavorited song"
   }
 }