+++ src/Game/Text/ScreenText.en.sjson.lua
@@ -1,0 +1,24 @@
+-- Used as a replacement for the Forget-Me-Not pin button prompt
+local order = {
+  "Id",
+  "DisplayName"
+}
+
+local newData = {
+  {
+    Id = "ModsNikkelMMusicMakerRandomizerFavoriteButton",
+    DisplayName = "{IP} SET AS FAVORITE"
+  },
+  {
+    Id = "ModsNikkelMMusicMakerRandomizerRemoveFavoriteButton",
+    DisplayName = "{IP} REMOVE FAVORITE"
+  }
+}
+
+local screenTextFile = rom.path.combine(rom.paths.Content, 'Game/Text/en/ScreenText.en.sjson')
+
+sjson.hook(screenTextFile, function(data)
+  for _, newCommand in ipairs(newData) do
+    table.insert(data.Texts, sjson.to_object(newCommand, order))
+  end
+end)