+++ src/Scripts/AudioLogic.lua
@@ -1,27 +1,51 @@
+local function IsSongEligibleForRandomizer(songName, randomizerType)
+	if songName == "Song_RandomSong" or songName == "Song_RandomSongFavorites" then
+		return false
+	end
+	if randomizerType == "Song_RandomSongTrack" then
+		return game.GameState.WorldUpgrades[songName] or config.AllSongs
+	elseif randomizerType == "Song_RandomSongFavoritesTrack" then
+		return game.GameState.WorldUpgrades[songName] and game.HasStoreItemPin(songName)
+	end
+	return false
+end
+
 modutil.mod.Path.Wrap("MusicianMusic", function(base, args)
 	-- If the chosen song is Song_RandomSong and the Music Maker not already playing something, choose a random track and save it in the currentRun to reset it only after a run ends
-	if args.TrackName == "Song_RandomSongTrack" then
+	-- Similarly, for the Song_RandomSongFavorites
+	if args.TrackName == "Song_RandomSongTrack" or args.TrackName == "Song_RandomSongFavoritesTrack" then
 		local availableTracks = {}
+
 		if game.GameState and game.GameState.WorldUpgrades then
+			-- Get all eligible tracks for the randomizer
 			for _, songName in ipairs(game.ScreenData.MusicPlayer.Songs) do
-				if songName ~= "Song_RandomSong" and (game.GameState.WorldUpgrades[songName] or config.AllSongs) then
+				if IsSongEligibleForRandomizer(songName, args.TrackName) then
 					table.insert(availableTracks, game.WorldUpgradeData[songName])
 				end
 			end
+
+			-- If no favorited songs exist, do not play anything
+			if args.TrackName == "Song_RandomSongFavorites" and #availableTracks == 0 then
+				args.TrackName = nil
+				game.GameState.MusicMakerRandomizerFriendlyPlayingSong = "Nothing... (No Favorites!)"
+				base(args)
+				return
+			end
+		else
+			args.TrackName = nil
+			game.GameState.MusicMakerRandomizerFriendlyPlayingSong = "Nothing.."
+			base(args)
+			return
 		end
 
-		-- Choose a random song from the available ones, if no other song was chosen yet in this "run"
-		if game.CurrentRun.MusicMakerRandomizerTrackName == nil then
-			local chosenTrack = availableTracks[game.RandomInt(1, #availableTracks)]
-			game.CurrentRun.MusicMakerRandomizerTrackName = chosenTrack.TrackName
-			-- To update the description of the Song_RandomSong to include the friendly name of the chosen track
-			game.GameState.MusicMakerRandomizerFriendlyPlayingSong = game.GetDisplayName({ Text = chosenTrack.Name })
-		end
+		-- Choose a random song from the available ones
+		local chosenTrack = availableTracks[game.RandomInt(1, #availableTracks)]
+		-- To update the description of the Song_RandomSong to include the friendly name of the chosen track
+		game.GameState.MusicMakerRandomizerFriendlyPlayingSong = game.GetDisplayName({ Text = chosenTrack.Name })
 
-		args.TrackName = game.CurrentRun.MusicMakerRandomizerTrackName
+		args.TrackName = chosenTrack.TrackName
 	else
-		-- If we are playing a different song, do not display its name in the description, as it won't get updated when the player chooses the random song, which may be confusing
-		GameState.MusicMakerRandomizerFriendlyPlayingSong = game.GetDisplayName({ Text = game.GameState.MusicPlayerSongName })
+		game.GameState.MusicMakerRandomizerFriendlyPlayingSong = game.GetDisplayName({ Text = game.GameState.MusicPlayerSongName})
 	end
 
 	base(args)