+++ src/Scripts/MusicPlayerData.lua
@@ -1,4 +1,5 @@
 table.insert(game.ScreenData.MusicPlayer.Songs, 1, "Song_RandomSong")
+table.insert(game.ScreenData.MusicPlayer.Songs, 2, "Song_RandomSongFavorites")
 
 local randomizerSong =
 {
@@ -9,7 +10,7 @@ local randomizerSong =
 
 		Cost =
 		{
-			CosmeticsPoints = 10,
+			CosmeticsPoints = 100,
 		},
 
 		PreRevealVoiceLines =
@@ -23,6 +24,19 @@ local randomizerSong =
 				},
 				{ Cue = "/VO/Melinoe_2356", Text = "Surprise!" },
 			},
+		},
+	},
+	Song_RandomSongFavorites = {
+		Name = "Song_RandomSongFavorites",
+		InheritFrom = { "DefaultSongItem" },
+		TrackName = "Song_RandomSongFavoritesTrack",
+
+		Cost =
+		{
+			CosmeticsPoints = 100,
+			GiftPoints = 1,
+			-- Custom resource that counts how many songs are currently favorited
+			ModsNikkelMMusicMakerRandomizerMusicPlayerFavoritesCount = 1,
 		},
 	},
 }