+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
-## v1.0.4
+## v2.0.0
 
 <!--Releasenotes start-->
-- Migrated the mod to the modern Thunderstore framework (Nexus will no longer receive updates).
+- NEW: Mark songs as favorite and have the Music Maker randomize only from favorites through a new song choice:
+![Randomize Music Maker Song Choice - Unlock favorites](https://github.com/user-attachments/assets/09733994-5346-44eb-988e-af1f039a8faa)
 <!--Releasenotes end-->
+
+## v1.0.4
+
+- Migrated the mod to the modern Thunderstore framework (Nexus will no longer receive updates).
 
 ## v1.0.3